Author: Charles Lepple <clepple+nut@gmail.com>

+++ server/netlist.c
@@ -52,7 +52,7 @@ static int tree_dump(st_tree_t *node, nut_ctype_t *client, const char *ups,
 		if (node->flags & ST_FLAG_RW) {
 			ret = sendback(client, "RW %s %s \"%s\"\n",
 				ups, node->var, node->val);
-		
+
 		} else {
 			ret = 1;	/* dummy */
 		}
@@ -240,10 +240,10 @@ static void list_ups(nut_ctype_t *client)
 			pconf_encode(utmp->desc, esc, sizeof(esc));
 			ret = sendback(client, "UPS %s \"%s\"\n",
 				utmp->name, esc);
-		
+
 		} else {
 			ret = sendback(client, "UPS %s \"Description unavailable\"\n",
-				 utmp->name);
+				utmp->name);
 		}
 
 		if (!ret)
@@ -253,7 +253,7 @@ static void list_ups(nut_ctype_t *client)
 	}
 
 	sendback(client, "END LIST UPS\n");
-}	
+}
 
 static void list_clients(nut_ctype_t *client, const char *upsname)
 {
@@ -285,7 +285,7 @@ static void list_clients(nut_ctype_t *client, const char *upsname)
 	sendback(client, "END LIST CLIENT %s\n", upsname);
 }
 
-void net_list(nut_ctype_t *client, int numarg, const char **arg)
+void net_list(nut_ctype_t *client, size_t numarg, const char **arg)
 {
 	if (numarg < 1) {
 		send_err(client, NUT_ERR_INVALID_ARGUMENT);
